Added pagination features to the controller and to the ejs files, as well as creating a include file for it

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../utils/fileFunctions');
 const { validationResult } = require('express-validator');
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -133,12 +134,14 @@ exports.postAddProduct = (req, res, next) => {
 
 exports.postEditProduct = (req, res, next) => {
   const { productId, title, price, description } = req.body;
-  const { image } = req.file;
-  /*  console.log(productId);
-  console.log(req.user._id.toString()); */
+  const image = req.file;
+
   const errors = validationResult(req);
 
+ 
+
   if (!errors.isEmpty()) {
+    console.log('errors');
     const { value, msg, param } = errors.array()[0];
     return res.status(422).render('admin/edit-product', {
       docTitle: 'Add Product',
@@ -158,24 +161,29 @@ exports.postEditProduct = (req, res, next) => {
 
   Product.findById(productId)
     .then((product) => {
+      console.log('made it into the product find');
       if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/');
+      }
+      if (image) {
+        console.log('imag exist');
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
       }
       product.title = title;
       product.price = price;
       product.price = price;
-      product.imageUrl = image.path || null;
       product.description = description;
 
       return product
         .save()
         .then((result) => {
-          console.log('Updated');
           res.redirect('/admin/products');
         })
         .catch((err) => console.log(err));
     })
     .catch((err) => {
+      console.log(err);
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
@@ -184,8 +192,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  /*   findByIdAndRemove */
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('Deleted');
       res.redirect('/admin/products');
